requirements.txt
appnope==0.1.0
backcall==0.1.0
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
dateparser==0.7.1
decorator==4.4.0
fake-useragent==0.1.11
idna==2.8
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.15.1
lxml==4.4.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2019.2
regex==2019.6.8
requests==2.22.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.2
traitlets==4.3.2
tzlocal==2.0.0
urllib3==1.25.3
wcwidth==0.1.7
